refactor: reorganize router configuration
